requirements.txt
dnspython==2.4.2
Faker==20.1.0
mongo==0.2.0
pymongo==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
